[Refactor] Move filesystem provider bootstrap into module bootstrap
