+++ be/src/fs/CMakeLists.txt
@@ -38,7 +38,6 @@ ADD_BE_LIB(FSCore
 target_link_libraries(FSCore PUBLIC IO)
 
 set(EXEC_FILES
-    fs_provider_bootstrap.cpp
     fs_options_helper.cpp
     fs_s3.cpp
     azure/azblob_uri.cpp
