+++ be/src/service/starrocks_main.cpp
@@ -63,8 +63,8 @@
 #include "exec/jdbc/jdbc_driver_manager.h"
 #include "exec/pipeline/query_context.h"
 #include "formats/orc/lzo_decompressor_registration.h"
-#include "fs/fs_provider_bootstrap.h"
 #include "fs/fs_s3.h"
+#include "module/fs_provider_bootstrap.h"
 #include "platform/aws/aws_sdk_guard.h"
 #include "platform/path_rw.h"
 #include "platform/store_path.h"
