+++ be/test/testutil/init_test_env.h
@@ -43,8 +43,8 @@
 #include "exec/pipeline/driver_queue_factory.h"
 #include "exec/pipeline/primitives/pipeline_metrics.h"
 #include "exec/pipeline/query_context.h"
-#include "fs/fs_provider_bootstrap.h"
 #include "gtest/gtest.h"
+#include "module/fs_provider_bootstrap.h"
 #include "orchestration/orchestration_env.h"
 #include "platform/platform_env.h"
 #include "platform/store_path.h"
