+++ build-support/check_be_module_boundaries.py
@@ -44,6 +44,7 @@ INTERNAL_INCLUDE_PREFIXES = {
     "gutil",
     "http",
     "io",
+    "module",
     "platform",
     "orchestration",
     "runtime",
